// packages/core/src/types.ts
export type ImageType = {
  id: string
  captionId: string
  imageCaption: string
  imageUrl: string
  tag: string
  score: number | string
}

export type ImageCacheResponse = {
  images: ImageType[]
}

export type CreateEventArgs = {
  type: string
  agent: string
  speaker: string
  text: string
  client: string
  channel: string
}

export type GetEventArgs = {
  type: string
  agent: string
  speaker: string
  client: string
  channel: string
  maxCount: number
}

export type EngineContext = {
  completion: (
    body: ModelCompletionOpts
  ) => Promise<string | OpenAIResultChoice | undefined>
  getCurrentGameState: () => Record<string, unknown>
  setCurrentGameState: (state: Record<string, unknown>) => void
  updateCurrentGameState: (update: Record<string, unknown>) => void
  enkiCompletion: (
    taskName: string,
    inputs: string[] | string
  ) => Promise<{ outputs: string[] }>
  huggingface: (
    model: string,
    request: string
  ) => Promise<{ error?: unknown; [key: string]: unknown }>
  runSpell: (
    flattenedInputs: Record<string, any>,
    spellId: string,
    state: Record<string, any>
  ) => Record<string, any>
  readFromImageCache: (
    caption: string,
    cacheTag?: string,
    topK?: number
  ) => Promise<ImageCacheResponse>
  processCode: (
    code: unknown,
    inputs: ThothWorkerInputs,
    data: Record<string, any>,
    state: Record<string, any>
  ) => any | void
  getEvent: (args: GetEventArgs) => Promise<string | string[] | null>
  storeEvent: (args: CreateEventArgs) => Promise<any>
  getWikipediaSummary: (keyword: string) => Promise<Record<string, any> | null>
}
// ...
export type ThothNode = Node & {
  inspector: Inspector
  display: (content: string) => void
  outputs: { name: string; [key: string]: unknown }[]
  category?: string
  deprecated?: boolean
  displayName?: string
  info: string
  subscription: Function
  console: ThothConsole
}

export type ModuleType = {
  id: string
  name: string
  data: GraphData
  createdAt: number
  updatedAt: number
}

export type ModelCompletionOpts = {
  model?: string
  prompt?: string
  maxTokens?: number
  temperature?: number
  topP?: number
  n?: number
  stream?: boolean
  logprobs?: number
  echo?: boolean
  stop?: string | string[]
  presencePenalty?: number
  frequencyPenalty?: number
  bestOf?: number
  user?: string
  logitBias?: { [token: string]: number }
}

export type OpenAIResultChoice = {
  text: string
  index: number
  logprobs: number[]
  top_logprobs: any[]
  text_offset: number[]
}
// ...
export type NodeData = ReteNodeData & {
  fewshot?: string
  display: Function
  error?: boolean
  console: ThothConsole
}

// export type Node = {
//   id: number,
//   data: NodeData,
//   name: string,
//   inputs: NodeOutputs,
//   outputs?: NodeOutputs,
//   position: number[]
// }

// export type Spell = {
//   id: string
//   nodes: Record<number, Node>
// }

export type ThothReteInput = {
  type: TaskOutputTypes
  outputData: unknown
  task: ThothTask
  key: string
}

export type TaskOutput = {
  type: TaskOutputTypes
  task: ThothTask
  key: string
}

export type ModuleWorkerOutput = WorkerOutputs & {
  [key: string]: any
}

export type ThothWorkerInput = string | unknown | unknown[]
export type ThothWorkerInputs = { [key: string]: ThothWorkerInput[] }
export type ThothWorkerOutputs = WorkerOutputs & {
  [key: string]: TaskOutput
}